fix: on creation, update or removal of file that defines an endpoint regenerate the correct endpoint instead of reloading all of them
move also the route generation to routeUtils
